:sparkles: integrate version src to one dirctory
