+++ Cargo.toml
@@ -22,6 +22,7 @@ flate2 = "1.0.22"
 tar = "0.4.38"
 dirs = "1.0.4"
 pathdiff = "0.2.1"
+which = "4.2.4"
 tempfile = "3.2.0"
 num_cpus = "1.13.1"
 colored = "1.6.1"
