+++ src/commands/use.rs
@@ -1,23 +1,23 @@
 use super::{Command, Config, ConfigError};
-use crate::alias::{self, Alias};
-use crate::decorized::Decorized;
+use crate::decorized::{color::Color, Decorized};
 use crate::symlink;
+use crate::version;
+use crate::version::Alias;
 use crate::version::Version;
-use crate::version_file::{self, VersionFile};
 use clap;
 use colored::Colorize;
 use derive_more::Display;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use std::str::FromStr;
 use thiserror::Error;
 
 #[derive(clap::Parser, Debug)]
 pub struct Use {
     #[clap(name = "version | alias", help = "semantic version or alias name")]
-    version_name: Option<VersionName>,
+    version_name: Option<RequestVersion>,
 
     #[clap(flatten)]
-    version_file: VersionFile,
+    version_file: version::File,
 
     /// Don't output a message to stdout
     #[clap(long)]
@@ -25,9 +25,10 @@ pub struct Use {
 }
 
 #[derive(Debug, Display)]
-enum VersionName {
+enum RequestVersion {
     Version(Version),
     Alias(Alias),
+    System,
 }
 
 #[derive(Error, Debug)]
@@ -39,13 +40,16 @@ pub enum Error {
     NoMultiShellPath(#[from] ConfigError),
 
     #[error(transparent)]
-    NotFoundAlias(#[from] alias::Error),
+    NotFoundAlias(#[from] version::alias::Error),
 
     #[error("Can't detect a version: {0}")]
-    NoVersionFromFile(#[from] version_file::Error),
+    NoVersionFromFile(#[from] version::file::Error),
 
     #[error("Can't find installed version '{0}', specified by '{1}'")]
     NotInstalledFromFile(Version, PathBuf),
+
+    #[error("Can't find a system version")]
+    NoSystemVersion,
 }
 
 macro_rules! outln {
@@ -61,10 +65,10 @@ impl Command for Use {
     fn run(&self, config: &Config) -> Result<(), Error> {
         let request_version = match &self.version_name {
             Some(version_name) => match version_name {
-                VersionName::Version(version) => config
+                RequestVersion::Version(version) => config
                     .latest_local_version_included_in(version)
                     .ok_or(Error::NotInstalled(*version))?,
-                VersionName::Alias(alias) => {
+                RequestVersion::Alias(alias) => {
                     let (_, version) = alias.resolve(config.aliases_dir())?;
                     outln!(
                         !self.quiet,
@@ -74,10 +78,21 @@ impl Command for Use {
                     );
                     version
                 }
+                RequestVersion::System => {
+                    let system_path = version::system::path().ok_or(Error::NoSystemVersion)?;
+                    replace_multishell_path(system_path, config)?;
+
+                    outln!(
+                        !self.quiet,
+                        "Using {} PHP",
+                        "system".color(<Version as Decorized>::Color::color())
+                    );
+                    return Ok(());
+                }
             },
             None => {
                 let info = match self.version_file.get_version_info() {
-                    Err(version_file::Error::NoVersionFile(_)) if self.quiet => return Ok(()),
+                    Err(version::file::Error::NoVersionFile(_)) if self.quiet => return Ok(()),
                     other => other,
                 }?;
                 outln!(
@@ -92,33 +107,34 @@ impl Command for Use {
             }
         };
 
-        let multishell_path = config.multishell_path()?;
-        let is_used_yet = multishell_path.exists();
         let version_dir = config.versions_dir().join(request_version.to_string());
-
-        symlink::remove(multishell_path).expect("Can't remove symlink!");
-        symlink::link(version_dir.join("bin"), multishell_path).expect("Can't create symlink!");
+        replace_multishell_path(version_dir.join("bin"), config)?;
 
         outln!(
             !self.quiet,
             "Using {}",
             request_version.decorized_with_prefix()
         );
-        if !is_used_yet {
-            outln!(
-                !self.quiet,
-                "{}: Need to type `rehash` in this shell if you are using zsh (only first time)",
-                "warning".yellow().bold()
-            );
-        }
         Ok(())
     }
 }
 
-impl FromStr for VersionName {
+fn replace_multishell_path(new_path: impl AsRef<Path>, config: &Config) -> Result<(), Error> {
+    let multishell_path = config.multishell_path()?;
+    symlink::remove(multishell_path).expect("Can't remove symlink!");
+    symlink::link(new_path, multishell_path).expect("Can't create symlink!");
+    Ok(())
+}
+
+impl FromStr for RequestVersion {
     type Err = Error;
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(s.parse::<Version>()
-            .map_or(Self::Alias(s.parse().unwrap()), Self::Version))
+        if s == "system" {
+            Ok(Self::System)
+        } else {
+            s.parse::<Version>()
+                .map(Self::Version)
+                .or_else(|_| Ok(Self::Alias(s.parse().unwrap())))
+        }
     }
 }
