+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub mod alias;
 pub mod cli;
 pub mod commands;
 pub mod config;
@@ -8,4 +7,3 @@ pub mod release;
 pub mod shell;
 pub mod symlink;
 pub mod version;
-pub mod version_file;
